+++ public/js/p3/widget/AdvancedSearchFields.js
@@ -714,7 +714,7 @@ define([], function () {
         field: 'accession', type: 'str', facet: false, facet_hidden: true, search: true
       },
       {
-        field: 'annotation', type: 'str', facet: true, facet_hidden: false, search: true
+        field: 'annotation', type: 'str', facet: true, facet_hidden: true, search: true
       },
       {
         field: 'alt_locus_tag', type: 'str', facet: false, facet_hidden: true, search: false
@@ -741,11 +741,11 @@ define([], function () {
         field: 'pathway_id', type: 'str', facet: false, facet_hidden: true, search: true
       },
       {
-        field: 'pathway_name', type: 'str', facet: true, facet_hidden: true, search: true
+        field: 'pathway_name', type: 'str', facet: true, facet_hidden: false, search: true
       },
       {
         field: 'pathway_class', type: 'str', facet: true, facet_hidden: false, search: true
-      },
+      }
     ],
     'subsystem': [
       {
@@ -779,9 +779,6 @@ define([], function () {
         field: 'subsystem_id', type: 'str', facet: false, facet_hidden: true, search: true
       },
       {
-        field: 'subsystem_name', type: 'str', facet: true, facet_hidden: false, search: true
-      },
-      {
         field: 'superclass', type: 'str', facet: true, facet_hidden: false, search: true
       },
       {
@@ -789,6 +786,9 @@ define([], function () {
       },
       {
         field: 'subclass', type: 'str', facet: true, facet_hidden: false, search: true
+      },
+      {
+        field: 'subsystem_name', type: 'str', facet: true, facet_hidden: false, search: true
       },
       {
         field: 'active', type: 'str', facet: true, facet_hidden: false, search: true