+++ public/js/p3/widget/TsvCsvFeatures.js
@@ -31,9 +31,14 @@ define([], function () {
     'gene_counts.tsv': { columnHeaders: true },
     'tpms.tsv': { columnHeaders: true },
 
-    // Subsystems service
+    // Subsystems Service
     '_subsystems.tsv': { columnHeaders: true },
     '_pathways.tsv': { columnHeaders: true },
+
+    // Subspecies Classification Service
+    'result.tsv': { columnHeaders: true },
+    'input.fasta.err': { columnHeaders: true },
+    'input.fasta.result': { columnHeaders: true },
   };
 
 });