+++ public/js/p3/widget/WorkspaceBrowser.js
@@ -52,7 +52,6 @@ define([
 
     startup: function () {
       var self = this;
-
       if (this._started) {
         return;
       }
@@ -266,7 +265,6 @@ define([
       }, false);
 
       // /START: ServicesGenomeGroups
-      // TODO: see if ServicesTooltipDialog works here now
       var dstContent = domConstruct.create('div', {});
       var viewGGServices = new TooltipDialog({
         content: dstContent,
@@ -1171,6 +1169,27 @@ define([
         }
       }, false);
 
+      this.browserHeader.addAction('ViewSubspeciesResult', 'fa icon-eye fa-2x', {
+        label: 'VIEW',
+        multiple: false,
+        validTypes: ['SubspeciesClassification'],
+        tooltip: 'View Result'
+      }, function (selection, container, button) {
+        console.log(selection);
+
+        var path;
+        selection[0].autoMeta.output_files.forEach(lang.hitch(this, function (meta_file_data) {
+          if (meta_file_data[0].includes('classification_report.html')) {
+            path = meta_file_data[0];
+          }
+        }));
+        if (path) {
+          Topic.publish('/navigate', { href: '/workspace' + encodePath(path) });
+        } else {
+          console.log('Error: could not find classification_report.html output file');
+        }
+      }, false);
+
       this.actionPanel.addAction('ViewAFA', 'fa icon-alignment fa-2x', {
         label: 'MSA',
         multiple: false,
@@ -1686,6 +1705,7 @@ define([
         'CodonTree': 'PhylogeneticTree',
         'PrimerDesign': 'PrimerDesign',
         'RNASeq': 'Rnaseq',
+        'SubspeciesClassification': 'SubspeciesClassification',
         'TaxonomicClassification': 'TaxonomicClassification',
         'TnSeq': 'Tnseq',
         'Variation': 'Variation'
@@ -1988,7 +2008,6 @@ define([
     },
 
     _setPathAttr: function (val) {
-
       // extract extra URL parameters
       var components = val.split('#');
       val = components[0];