+++ public/js/p3/widget/viewer/TSV_CSV.js
@@ -130,9 +130,10 @@ define([
       var downld = '<div><a href=' + this.url + '><i class="fa icon-download pull-left fa-2x"></i></a></div>';
 
       var ta_keyword = this.ta_keyword = new TextArea({
-        style: 'width:272px; min-height:20px; marginRight: 2.0em'
+        style: 'width:272px; min-height:20px; max-height:24px; marginRight: 2.0em'
       });
-      var label_keyword = domConstruct.create('label', { innerHTML: 'KEYWORDS   ' });
+      //
+      // var label_keyword = domConstruct.create('label', { innerHTML: 'KEYWORDS   ' });
 
       var columnFilter = domConstruct.create('span', {
         style: {
@@ -172,9 +173,39 @@ define([
       selector.set('options', items).reset();
 
       domConstruct.place(downld, this.filterPanel.containerNode, 'last');
-      domConstruct.place(label_keyword, this.filterPanel.containerNode, 'last');
+
+      var filterType = domConstruct.create('span', {
+        style: {
+          'float': 'right'
+        }
+      });
+      var filterSelect = new Select({
+        name: 'filterSelect',
+        style: {
+          width: '150px', marginRight: '2.0em'
+        },
+        options: [
+          { value: 'fuzzy', label: 'Substring Filter', selected: true },
+          { value: 'exact', label: 'Exact Filter' },
+          { value: 'range', label: 'Range Filter' },
+        ]
+      }, filterType);
+
+      domConstruct.place(filterSelect.domNode, this.filterPanel.containerNode, 'last');
       domConstruct.place(ta_keyword.domNode, this.filterPanel.containerNode, 'last');
       domConstruct.place(selector.domNode, this.filterPanel.containerNode, 'last');
+
+
+      var apply_filter = new Button({
+        label: 'Apply',
+        onClick: lang.hitch(this, function () {
+          var filter = {};
+          filter.keyword = ta_keyword.get('value');
+          filter.columnSelection = selector.get('value');
+          filter.type = filterSelect.get('value')
+          Topic.publish('applyKeywordFilter', filter);
+        })
+      });
 
       var btn_reset = new Button({
         label: 'Reset',
@@ -182,27 +213,11 @@ define([
 
           ta_keyword.set('value', '');
           selector.set('value', 'All Columns');
-          var filter = {};
-          filter.keyword = '';
-          filter.columnSelection = 'All Columns';
-
-          Topic.publish('applyKeywordFilter', filter);
+          Topic.publish('applyResetFilter');
         })
       });
 
-      var btn_submit = new Button({
-        label: 'Filter',
-        onClick: lang.hitch(this, function () {
-
-          var filter = {};
-
-          filter.keyword = ta_keyword.get('value');
-          filter.columnSelection = selector.get('value');
-
-          Topic.publish('applyKeywordFilter', filter);
-        })
-      });
-      domConstruct.place(btn_submit.domNode, this.filterPanel.containerNode, 'last');
+      domConstruct.place(apply_filter.domNode, this.filterPanel.containerNode, 'last');
       domConstruct.place(btn_reset.domNode, this.filterPanel.containerNode, 'last');
 
       // add button or checkbox for column headers if suffix is not known (user supplied data file)
@@ -267,23 +282,23 @@ define([
         max: 5000,
         validContainerTypes: ['csvFeature', '']
       },
-        function (selection, container) {
-          _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('selection', selection);
-          _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('containerType', this.containerType);
-          if (container && container.tsvGC.grid) {
-            _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('grid', container.tsvGC.grid);
-          }
+      function (selection, container) {
+        _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('selection', selection);
+        _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('containerType', this.containerType);
+        if (container && container.tsvGC.grid) {
+          _self.actionPanel._actions.CopySelection.options.tooltipDialog.set('grid', container.tsvGC.grid);
+        }
 
-          _self.actionPanel._actions.CopySelection.options.tooltipDialog.timeout(3500);
+        _self.actionPanel._actions.CopySelection.options.tooltipDialog.timeout(3500);
 
-          setTimeout(lang.hitch(this, function () {
-            popup.open({
-              popup: _self.actionPanel._actions.CopySelection.options.tooltipDialog,
-              around: _self.actionPanel._actions.CopySelection.button,
-              orient: ['below']
-            });
-          }), 10);
-        });
+        setTimeout(lang.hitch(this, function () {
+          popup.open({
+            popup: _self.actionPanel._actions.CopySelection.options.tooltipDialog,
+            around: _self.actionPanel._actions.CopySelection.button,
+            orient: ['below']
+          });
+        }), 10);
+      });
 
       var numColumns = Object.keys(data[0]).length;
       var checkFeatureIDs = [];