+++ public/worker/PathwayServiceWorker.js
@@ -1,4 +1,28 @@
 
+function load_data(text_data) {
+  var return_data = [];
+  var header = true;
+  var header_keys = null;
+  var idx = 0;
+  text_data.split('\n').forEach(function (line) {
+    if (header) {
+      header_keys = line.split('\t');
+      header = false;
+    }
+    else {
+      var new_data = {};
+      var l = line.split('\t');
+      header_keys.forEach(function (key, index) {
+        new_data[key] = l[index];
+      });
+      new_data['ec_index'] = idx + '';
+      return_data.push(new_data);
+      idx++;
+    }
+  });
+  return return_data;
+}
+
 function parseData_v2(data, primaryKey) {
   var data_dict = {};
   data.forEach(function (obj) {
@@ -78,6 +102,11 @@ onmessage = (msg) => {
       var key = payload.primaryKey;
       var parsed_data = parseData_v2(data, key);
       postMessage({ type: 'parsed_data', data: parsed_data });
+      break;
+    case 'load_data':
+      var text_data = payload.text_data;
+      var return_data = load_data(text_data);
+      postMessage({ type: 'loaded_data', data: return_data });
       break;
   }
 }